#1 add GPL-3.0 license for borrowed from `gr-satellites` code
