+++ SatsDecoder/systems/sonate2.py
@@ -1,6 +1,6 @@
 import construct
 
-from SatsDecoder.systems import ccsds, common
+from SatsDecoder.systems import ax25, ccsds, common
 
 __all__ = 'SonateProtocol',
 
@@ -8524,7 +8524,7 @@ class SonateProtocol:
 
     @staticmethod
     def get_sender_callsign(data):
-        return common.ax25_get_sender_callsign(data.header)
+        return ax25.get_sender_callsign(data.header)
 
     def recognize(self, bb):
         frame = ccsds.parse_space_packet(bb)
